bug fixes
fixed the weird text duplicatino issue, and also the server disconnects. learned what useCallback does and why we use ref. in this case we used a ref to represent our quill funciton because we wanted to memoize our quill, and our ref allowed us to prevent unecessary rendering (i think)